test: port challengeIFENotCanonical with the changes
1. Refactor test setup. Move stuff from ife.js back to the test file, and use global constants to represent
the general test setup values.
2. Make sures the tests passes with the new changes for spending condition and tx finalization
3. add test on whent outputGuardHandler fails

+++ plasma_framework/contracts/src/exits/payment/routers/PaymentInFlightExitRouterArgs.sol
@@ -67,7 +67,7 @@ library PaymentInFlightExitRouterArgs {
      */
     struct ChallengeCanonicityArgs {
         bytes inputTx;
-        uint16 inputUtxoPos;
+        uint256 inputUtxoPos;
         bytes inFlightTx;
         uint16 inFlightTxInputIndex;
         bytes competingTx;
